requirements.txt
discord.py
vaderSentiment
python-dotenv